feat: add write callbacks to characteristic
This worked first time and I just wanted to share that with everyone!

I'm really proud of my solution too. In order to allow you to write callbacks for multiple characterstics in different services, I just stored pointers to a struct containing all three so they can be searched for. This avoided me having to write a convoluted OnCharacteristicWrite method which would become a huge mess any time I try to add a new characteristic.